move RTE to its own file
